# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "FyAPI",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/api/v1",
    "paths": {
...
        "/public/ticket/id/{id}": {
            "get": {
                "description": "Get public ticket by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ticket"
                ],
                "summary": "Get public by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully get public ticket by id",
                        "schema": {
                            "$ref": "#/definitions/dto.APIResponse-dto_TicketGet"
                        }
                    },
                    "500": {
                        "description": "Failed to get public ticket by id",
                        "schema": {
                            "$ref": "#/definitions/dto.APIResponse-any"
                        }
                    }
                }
            }
        },
        "/public/ticket/{username}": {
            "get": {
                "description": "Get public tickets by username",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ticket"
                ],
                "summary": "Get public by username",
                "parameters": [
                    {
                        "type": "string",
                        "description": "username",
                        "name": "username",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully get public tickets by username",
                        "schema": {
                            "$ref": "#/definitions/dto.APIResponse-array_dto_TicketPublicGet"
                        }
                    },
                    "500": {
                        "description": "Failed to get public tickets by username",
                        "schema": {
                            "$ref": "#/definitions/dto.APIResponse-any"
                        }
                    }
                }
            }
        },
...
        "/ticket": {
            "get": {
                "description": "Get tickets",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ticket"
                ],
                "summary": "Get",
                "responses": {
                    "200": {
                        "description": "Successfully get tickets",
                        "schema": {
                            "$ref": "#/definitions/dto.APIResponse-array_dto_TicketGet"
                        }
                    },
                    "500": {
                        "description": "Failed to get tickets",
                        "schema": {
                            "$ref": "#/definitions/dto.APIResponse-any"
                        }
                    }
                }
            },
            "post": {
                "description": "Create new ticket",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ticket"
                ],
                "summary": "Create",
                "parameters": [
                    {
                        "type": "string",
                        "description": "title",
                        "name": "title",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "description",
                        "name": "description",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "description": "mode",
                        "name": "mode",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "quota",
                        "name": "quota",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "image",
                        "name": "image",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successfully create ticket",
                        "schema": {
                            "$ref": "#/definitions/dto.APIResponse-any"
                        }
                    },
                    "500": {
                        "description": "Failed to create ticket",
                        "schema": {
                            "$ref": "#/definitions/dto.APIResponse-any"
                        }
                    }
                }
            }
        },
        "/ticket/{id}": {
            "get": {
                "description": "Get ticket by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ticket"
                ],
                "summary": "Get by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully get ticket by id",
                        "schema": {
                            "$ref": "#/definitions/dto.APIResponse-dto_TicketGet"
                        }
                    },
                    "500": {
                        "description": "Failed to get ticket by id",
                        "schema": {
                            "$ref": "#/definitions/dto.APIResponse-any"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete ticket by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ticket"
                ],
                "summary": "Delete",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully delete ticket",
                        "schema": {
                            "$ref": "#/definitions/dto.APIResponse-any"
                        }
                    },
                    "500": {
                        "description": "Failed to delete ticket",
                        "schema": {
                            "$ref": "#/definitions/dto.APIResponse-any"
                        }
                    }
                }
            },
            "patch": {
                "description": "Update ticket by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ticket"
                ],
                "summary": "Update",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "title",
                        "name": "title",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "description",
                        "name": "description",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "description": "mode",
                        "name": "mode",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "quota",
                        "name": "quota",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "image",
                        "name": "image",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully update ticket",
                        "schema": {
                            "$ref": "#/definitions/dto.APIResponse-any"
                        }
                    },
                    "500": {
                        "description": "Failed to update ticket",
                        "schema": {
                            "$ref": "#/definitions/dto.APIResponse-any"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "dto.APIResponse-any": {
            "type": "object",
            "properties": {
                "data": {},
                "meta": {
                    "$ref": "#/definitions/dto.ApiMeta"
                }
            }
        },
...
        "dto.APIResponse-array_dto_TicketGet": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.TicketGet"
                    }
                },
                "meta": {
                    "$ref": "#/definitions/dto.ApiMeta"
                }
            }
        },
        "dto.APIResponse-array_dto_TicketPublicGet": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.TicketPublicGet"
                    }
                },
                "meta": {
                    "$ref": "#/definitions/dto.ApiMeta"
                }
            }
        },
...
        "dto.APIResponse-dto_TicketGet": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/dto.TicketGet"
                },
                "meta": {
                    "$ref": "#/definitions/dto.ApiMeta"
                }
            }
        },
...
        "dto.ApiMeta": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "request_id": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "dto.TicketGet": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "image": {
                    "type": "string"
                },
                "mode": {
                    "type": "boolean"
                },
                "quota": {
                    "type": "integer"
                },
                "registered_count": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "dto.TicketPublicGet": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "image": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
    }
}